conditionally copy the pkgin cache on existence

+++ hookit/hooks/build-cleanup.rb
@@ -22,12 +22,19 @@ execute "copy build into place" do
   command "cp -r #{BUILD_DIR}/* #{DEPLOY_DIR}"
 end
 
-# ensure the directory exists for pkgin cache
-directory "#{CACHE_DIR}/pkgin" do
-  recursive true
+if ::File.exist? "#{BUILD_DIR}/var/db/pkgin"
+
+  # ensure the directory exists for pkgin cache
+  directory "#{CACHE_DIR}/pkgin" do
+    recursive true
+  end
+
+  # copy (and remove) the pkgin cache & db for quick subsequent deploys
+  execute "stash pkgin packages into cache for quick access" do
+    command 'cp -r #{BUILD_DIR}/var/db/pkgin/* #{CACHE_DIR}/pkgin'
+  end
+
 end
 
-# copy (and remove) the pkgin cache & db for quick subsequent deploys
-execute "stash pkgin packages into cache for quick access" do
-  command 'cp -r #{BUILD_DIR}/var/db/pkgin/* #{CACHE_DIR}/pkgin'
-end
+# temporarily exit to inspect
+exit 1
